Add files via upload
Homework assignment for module 15 in my Skillbox JavaScript course. A react app that lets adding comments: a user should give their name and comment text, and both is added along with the current time. Each comment has a ‘delete’ button next to it, that allows deleting.
Note: if you want to re-assemble with webpack, remember to put the index.js file into src/js folder and subfolder, or webpack won't see it. 
Or just unpack the archive with the whole lot.